break down page file
